Improved Radar 2/pyproject.toml
[project]
name = "soloist-agent"
version = "0.1.0"
description = "Autonomous freelance gig discovery engine"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "openai>=1.60.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "chromadb>=0.5.0",
    "playwright>=1.42.0",
    "playwright-stealth>=1.0.6",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "markdown2>=2.5.0",
    "weasyprint>=62.0",
    "sqlalchemy>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "aiofiles>=24.1.0",
]

[project.scripts]
soloist = "app.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
